--- common/IWriterRegistry.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from abc import ABC, abstractmethod
-
-class IWriterRegistry(ABC):
-    
-    @abstractmethod
-    def getId(self,*args):
-        pass
