--- common/records/Serializer.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-class Serializer:
-    def __init__(self,string_byte):
-        self.string_byte=string_byte
-        
-    def put(self,value):
-           self.string_byte.append(value+"; ")
