+++ src/Record.py
@@ -1,5 +1,11 @@
 # -*- coding: utf-8 -*-
 
+class Serializer:
+    def __init__(self,string_byte):
+        self.string_byte=string_byte
+        
+    def put(self,value):
+           self.string_byte.append(value+"; ")
 class DummyRecord:
     
     def __init__(self, foo_parameter, bar_parameter):
